scrollbar, dialog styling, relative width, drop-shadow, filter-blur
figured out how to change image-div's sizing into vw instead of px without it messing up the js-code of all the image-div images still staying inside the viewport, and added a drop-shadow for them. Also added styling to the scrollbar and did some detailing with the dialog-divs. blurred out the header slightly